requirements.txt
mysql-connector-python
prometheus-client